SearchKeyword as interface #1053
