+++ commercetools-sdk-base/src/main/java/io/sphere/sdk/search/SearchKeyword.java
@@ -1,36 +1,22 @@
 package io.sphere.sdk.search;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.search.tokenizer.SuggestTokenizer;
 
 import javax.annotation.Nullable;
 
-public final class SearchKeyword extends Base {
-    private final String text;
+@JsonDeserialize(as = SearchKeywordImpl.class)
+public interface SearchKeyword {
     @Nullable
-    private final SuggestTokenizer suggestTokenizer;
+    SuggestTokenizer getSuggestTokenizer();
 
-    @JsonCreator
-    private SearchKeyword(final String text, @Nullable final SuggestTokenizer suggestTokenizer) {
-        this.suggestTokenizer = suggestTokenizer;
-        this.text = text;
-    }
+    String getText();
 
-    @Nullable
-    public SuggestTokenizer getSuggestTokenizer() {
-        return suggestTokenizer;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public static SearchKeyword of(final String text) {
+    static SearchKeyword of(final String text) {
         return of(text, null);
     }
 
-    public static SearchKeyword of(final String text, final SuggestTokenizer suggestTokenizer) {
-        return new SearchKeyword(text, suggestTokenizer);
+    static SearchKeyword of(final String text, final SuggestTokenizer suggestTokenizer) {
+        return new SearchKeywordImpl(text, suggestTokenizer);
     }
 }
